+++ src/Hive.IO/Hive.IO/Plots/AmrPlotBase.cs
@@ -2,7 +2,7 @@
 using Grasshopper.GUI;
 using Grasshopper.GUI.Canvas;
 using Grasshopper.Kernel;
-using Hive.IO.DataHandling;
+using Hive.IO.Results;
 
 namespace Hive.IO.Plots
 {
