+++ src/Hive.IO/Hive.IO/Plots/DemandMonthlyNormalizedPlot.cs
@@ -1,5 +1,5 @@
 ﻿using System.Linq;
-using Hive.IO.DataHandling;
+using Hive.IO.Results;
 using OxyPlot;
 using OxyPlot.Axes;
 using OxyPlot.Series;
