+++ src/Hive.IO/Hive.IO/Plots/EnergyBalancePlot.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using Grasshopper.GUI;
 using Grasshopper.GUI.Canvas;
 using Grasshopper.Kernel;
-using Hive.IO.DataHandling;
+using Hive.IO.Results;
 
 namespace Hive.IO.Plots
 {
