+++ src/Hive.IO/Hive.IO/Plots/IVisualizerPlot.cs
@@ -1,7 +1,7 @@
 ﻿using System.Drawing;
 using Grasshopper.GUI;
 using Grasshopper.GUI.Canvas;
-using Hive.IO.DataHandling;
+using Hive.IO.Results;
 
 namespace Hive.IO.Plots
 {
