+++ src/Hive.IO/Hive.IO/Plots/OxyPlotBase.cs
@@ -1,7 +1,7 @@
 ﻿using System.Drawing;
 using Grasshopper.GUI;
 using Grasshopper.GUI.Canvas;
-using Hive.IO.DataHandling;
+using Hive.IO.Results;
 using OxyPlot;
 using OxyPlot.WindowsForms;
